refactor: Split ty.rs into files
